5 acceptance test

+++ app/controllers/questions_controller.rb
@@ -1,4 +1,7 @@
 class QuestionsController < ApplicationController
+  before_action :question, only: %i[show edit]
+  before_action :authenticate_user!, except: %i[index show]
+
   def index
     @questions = Question.all
   end
@@ -7,15 +10,16 @@ class QuestionsController < ApplicationController
   end
 
   def new
+    @question = current_user.questions.new
   end
 
   def edit
   end
 
   def create
-    @question = Question.new(question_params)
-    if @question.save
-      redirect_to @question
+    @question = current_user.questions.new(question_params)
+    if question.save
+      redirect_to question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -23,14 +27,14 @@ class QuestionsController < ApplicationController
 
   def update
     if question.update(question_params)
-      redirect_to @question
+      redirect_to question
     else
       render :edit
     end
   end
 
   def destroy
-    question.destroy
+    question.destroy if current_user.author?(question)
     redirect_to questions_path
   end
